feat: rewrite in TypeScript
BREAKING CHANGE: main export is now named

Allows for a single source of truth and multiple bundle targets.
Support both ESM and CJS (through UMD).
