+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "compare-versions",
-  "version": "4.1.4",
+  "version": "4.1.3",
   "description": "Compare semver version strings to find greater, equal or lesser.",
   "repository": {
     "type": "git",
@@ -20,19 +20,24 @@
     "node"
   ],
   "scripts": {
-    "test": "c8 --reporter=lcov mocha"
+    "build": "npm run build:esm && npm run build:umd",
+    "build:esm": "tsc --module esnext --target es2017 --outDir lib/esm",
+    "build:umd": "rollup lib/esm/index.js --format umd --name compareVersions --sourcemap -o lib/umd/index.js",
+    "test": "c8 --reporter=lcov mocha -r ts-node/register test/**/*.ts"
   },
-  "main": "index.js",
-  "module": "index.mjs",
-  "types": "index.d.ts",
+  "main": "./lib/umd/index.js",
+  "module": "./lib/esm/index.js",
+  "types": "./lib/esm/index.d.ts",
   "sideEffects": false,
   "files": [
-    "index.js",
-    "index.mjs",
-    "index.d.ts"
+    "lib"
   ],
   "devDependencies": {
+    "@types/mocha": "^9.1.0",
     "c8": "^7.10.0",
-    "mocha": "^10.0.0"
+    "mocha": "^10.0.0",
+    "rollup": "^2.78.1",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3"
   }
 }
