+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+	"compilerOptions": {
+		"declaration": true,
+		"esModuleInterop": true,
+		"module": "UMD",
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"outDir": ".",
+		"sourceMap": true,
+		"strict": true,
+		"target": "ES5"
+	},
+	"include": [
+		"src/index.ts"
+	],
+	"exclude": [
+		"./src/test/**/*",
+		"node_modules"
+	]
+}
